01_Middle_Data/01_ConvertScript_V4/requirements.txt
PyPDF2
pdf2docx
pdf2image
aspose.words
bs4
tqdm